Added Wattpad, Tapas, Aurelia, Gumroad icons (#763)
* Add files via upload

* Add Wattpad and Tapas

* Add Aurelia icon

* Add Aurelia to json

Chose "ED2B88" color based on the buttons on the site; since the logo and most design elements have gradients it's tricky-feeling to choose one color. I think this choice is backed up though, because the "a" in the text logo is "ED2C88", just a tad brigher than the color I chose, and the "a" is the brightest letter in the logo. Let me know if you think otherwise.

* Add Gumroad icon

* Added Gumroad to data

Green and orange display prominently in the logo, but green is used for the website buttons and seems to be more prominent in overall website design and marketing. I chose the base color for their buttons- not the gradient slight lighter than/darker than variants.

* Remove rogue fill on Gumroad

Dagnabit, why'd they put it at the end? XD

* Remove extra line that appeared

* Delete gumroad.svg

* Replace gumroad

To fix extra line being added and rogue fill

* Set to pages for icons, not direct icon links

For Tapas, Gumroad, Aurelia, and Wattpad

* Go to id assets for Wattpad