Update CONTRIBUTING.md
Add svg-grabber with the "Website headers" option for where you can find brand logos. svg-grabber is a tool (for Google Chrome) which I recently discovered, it quickly shows you all the available SVG files on a website in order. So most of the time the company logo will be listed somewhere at the top. This might make the process of finding the SVG easier for some.