Update colors of Gitter, Ghost and Evernote (#848)
* Update Gitter color

Update the color for the Gitter logo. Since it is hard to find an official
source for the color, but pink seems to be more or less their brand color,
I used the pink found at https://brandcolors.net/b/gitter

Closes #820

* Update Ghost color

Updated the color of the Ghost icon based on their (updated) styleguide

Closes #824

* Update Evernote color

Updated the color for the Evernote icon. Although the green is more
iconic, their brand guidelines specify explicitely that you should not
change the color of the logo.

Closes #813

* Update Evernote color

Updated the Evernote color based on feedback by @birjolaxew using the
primary brand color rather then the icon color