Publish 1.6.0 (#849)
* remove aim (#827)

* update etsy color (#828)

* update feedly color/source (#829)

* update freeCodeCamp (#830)

* update garmin color (#831)

* update gatsby color (#832)

* update geocaching (#833)

* update git color (#834)

* update gauges color (#835)

* Update Viber (#836)

* Update Viber SVG

* Update Viber color and source

* Version bump

* Version bump (removal)

* Remove Gratipay icon (#847)

The project stopped in 2017, see:
https://gratipay.news/the-end-cbfba8f50981

* Remove Geeklist icon (#845)

The website seems to have stopped

* Add Debian logo (#844)

Add the Debian logo taken from the provided link, the color is extracted
from the SVG found there.

* Update colors of Gitter, Ghost and Evernote (#848)

* Update Gitter color

Update the color for the Gitter logo. Since it is hard to find an official
source for the color, but pink seems to be more or less their brand color,
I used the pink found at https://brandcolors.net/b/gitter

Closes #820

* Update Ghost color

Updated the color of the Ghost icon based on their (updated) styleguide

Closes #824

* Update Evernote color

Updated the color for the Evernote icon. Although the green is more
iconic, their brand guidelines specify explicitely that you should not
change the color of the logo.

Closes #813

* Update Evernote color

Updated the Evernote color based on feedback by @birjolaxew using the
primary brand color rather then the icon color

* Add Mail.Ru icon (#846)

* Add Mail.Ru icon

Added the icon for Mail.Ru taken from the source provided, using the
orange RGB found there as well because of the color of the sites favicon.

* Update Mail.Ru color

Updated the color for Mail.Ru based on feedback by @birjolaxew using the
primary brand color instead of the icon color.

* Version bump