Add Ko-fi icon (#869)
* Add Ko-fi icon

Add icon for Ko-fi, a service similar to Patreon, only without recurring
donations. Icon is based on the SVG found on their About page (https://ko-fi.com/home/about).

* Add Ko-fi details

Update the simple-icons.json file to include Ko-fi. The source links to
the page where the SVG was found, the color is taken from this SVG.

* Fix indentation for Ko-fi data